+++ cmd/main.go
@@ -65,12 +65,11 @@ func main() {
 	defer restore()
 
 	c := gopty.NewController(m, os.Stdin, os.Stdout)
-	listenResize(ctx, m.ResizeAll)
+	listenResize(ctx, c.ResizeAll)
 	listenTerm(ctx, c.Shutdown)
 
 	go c.Run()
-	m.WaitAll()
-	c.Cleanup()
+	c.Wait()
 }
 
 func parseEntries(path, filter string) ([]gopty.Entry, error) {